GTM context now throws error if used outside of provider
- also moved GTM provider from AppPageContent
to _app.tsx, so it wraps the entire component tree
