--- project-base/storefront/gtm/context/useGtmContext.ts
@@ -1,13 +1,0 @@
-import { GtmContext, GtmContextType } from './GtmProvider';
-import { useContext } from 'react';
-
-export const useGtmContext = (): GtmContextType => {
-    const context = useContext(GtmContext);
-
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-    if (!context) {
-        throw new Error('useGtmContext must be used within a GtmProvider');
-    }
-
-    return context;
-};
