+++ project-base/storefront/gtm/utils/pageViewEvents/useGtmProductDetailViewEvent.ts
@@ -1,7 +1,7 @@
 import { useDomainConfig } from 'components/providers/DomainConfigProvider';
 import { TypeMainVariantDetailFragment } from 'graphql/requests/products/fragments/MainVariantDetailFragment.generated';
 import { TypeProductDetailFragment } from 'graphql/requests/products/fragments/ProductDetailFragment.generated';
-import { useGtmContext } from 'gtm/context/useGtmContext';
+import { useGtmContext } from 'gtm/context/GtmProvider';
 import { getGtmProductDetailViewEvent } from 'gtm/factories/getGtmProductDetailViewEvent';
 import { gtmSafePushEvent } from 'gtm/utils/gtmSafePushEvent';
 import { useEffect, useRef } from 'react';
